backend/python-api/requirements.txt
# API Dependencies
requests>=2.32.3
python-dotenv>=1.1.0

# Database Dependencies
pymysql>=1.1.0
sqlalchemy>=2.0.0

# Web Framework
fastapi>=0.110.0
uvicorn>=0.27.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Testing
pytest>=8.0.0